[reorg] Move LigatureSubst to new layout.
https://github.com/harfbuzz/harfbuzz/commit/403feb380
