+++ src/hb/ot/layout/GSUB/mod.rs
@@ -1,5 +1,8 @@
 mod alternate_set;
 mod alternate_subst;
+mod ligature;
+mod ligature_set;
+mod ligature_subst;
 mod multi_subst;
 mod sequence;
 mod single_subst;
