+++ src/hb/ot_layout_gsub_table.rs
@@ -10,100 +10,6 @@ use super::ot_layout_gsubgpos::*;
 use super::ot_shape_plan::hb_ot_shape_plan_t;
 use OT::hb_ot_apply_context_t;
 
-impl WouldApply for Ligature<'_> {
-    fn would_apply(&self, ctx: &WouldApplyContext) -> bool {
-        ctx.glyphs.len() == usize::from(self.components.len()) + 1
-            && self
-                .components
-                .into_iter()
-                .enumerate()
-                .all(|(i, comp)| ctx.glyphs[i + 1] == comp)
-    }
-}
-
-impl Apply for Ligature<'_> {
-    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
-        // Special-case to make it in-place and not consider this
-        // as a "ligated" substitution.
-        if self.components.is_empty() {
-            ctx.replace_glyph(self.glyph);
-            Some(())
-        } else {
-            let f = |glyph, num_items| {
-                let index = self.components.len() - num_items;
-                let value = self.components.get(index).unwrap();
-                match_glyph(glyph, value.0)
-            };
-
-            let mut match_end = 0;
-            let mut match_positions = [0; MAX_CONTEXT_LENGTH];
-            let mut total_component_count = 0;
-
-            if !match_input(
-                ctx,
-                self.components.len(),
-                &f,
-                &mut match_end,
-                &mut match_positions,
-                Some(&mut total_component_count),
-            ) {
-                ctx.buffer
-                    .unsafe_to_concat(Some(ctx.buffer.idx), Some(match_end));
-                return None;
-            }
-
-            let count = usize::from(self.components.len()) + 1;
-            ligate_input(
-                ctx,
-                count,
-                &match_positions,
-                match_end,
-                total_component_count,
-                self.glyph,
-            );
-            return Some(());
-        }
-    }
-}
-
-impl WouldApply for LigatureSet<'_> {
-    fn would_apply(&self, ctx: &WouldApplyContext) -> bool {
-        self.into_iter().any(|lig| lig.would_apply(ctx))
-    }
-}
-
-impl Apply for LigatureSet<'_> {
-    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
-        for lig in self.into_iter() {
-            if lig.apply(ctx).is_some() {
-                return Some(());
-            }
-        }
-        None
-    }
-}
-
-// LigatureSubstFormat1::would_apply
-impl WouldApply for LigatureSubstitution<'_> {
-    fn would_apply(&self, ctx: &WouldApplyContext) -> bool {
-        self.coverage
-            .get(ctx.glyphs[0])
-            .and_then(|index| self.ligature_sets.get(index))
-            .map_or(false, |set| set.would_apply(ctx))
-    }
-}
-
-// LigatureSubstFormat1::apply
-impl Apply for LigatureSubstitution<'_> {
-    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
-        let glyph = ctx.buffer.cur(0).as_glyph();
-        self.coverage
-            .get(glyph)
-            .and_then(|index| self.ligature_sets.get(index))
-            .and_then(|set| set.apply(ctx))
-    }
-}
-
 // ReverseChainSingleSubstFormat1::would_apply
 impl WouldApply for ReverseChainSingleSubstitution<'_> {
     fn would_apply(&self, ctx: &WouldApplyContext) -> bool {
